+++ src/initialize.js
@@ -17,6 +17,7 @@ const { getAllTenantIds } = require("./shared/cdsHelper");
 const { EventProcessingStatus } = require("./constants");
 const distributedLock = require("./shared/distributedLock");
 const EventQueueError = require("./EventQueueError");
+const { initMetrics } = require("./shared/openTelemetry");
 
 const readFileAsync = promisify(fs.readFile);
 
@@ -49,6 +50,7 @@ const CONFIG_VARS = [
   ["disableProcessingOfSuspendedTenants", true],
   ["namespace", "default"],
   ["processingNamespaces", ["default"]],
+  ["collectEventQueueMetrics", false],
 ];
 
 /**
@@ -78,6 +80,7 @@ const CONFIG_VARS = [
  * @param {string} [options.crashOnRedisUnavailable=true] - If enabled an error is thrown if the redis connection check is not successful
  * @param {string} [options.namespace=default] - Default namespace in which events are published
  * @param {string} [options.processingNamespaces=[default]] - Namespaces which the application processes
+ * @param {boolean} [options.collectEventQueueMetrics=false] - Enable collection of event queue metrics (pending/inProgress counters) stored in Redis and exposed via OpenTelemetry gauges.
  */
 const initialize = async (options = {}) => {
   if (config.initialized) {
@@ -125,6 +128,7 @@ const initialize = async (options = {}) => {
     runInterval: config.runInterval,
     useAsCAPQueue: config.useAsCAPQueue,
   });
+  initMetrics();
   resolveFn();
 };