+++ test-integration/e2e-redis.test.js
@@ -75,6 +75,7 @@ describe("end-to-end", () => {
       isEventQueueActive: true,
     });
     eventQueue.config.redisEnabled = true;
+    eventQueue.config.collectEventQueueMetrics = true;
 
     cds.emit("connect", await cds.connect.to("db"));
   });
@@ -82,7 +83,6 @@ describe("end-to-end", () => {
   beforeEach(async () => {
     await DELETE.from("sap.eventqueue.Lock");
     await DELETE.from("sap.eventqueue.Event");
-    await DELETE.from("cds.outbox.Messages");
     jest.clearAllMocks();
     redisMock.clearState();
     redisMock.clearTestState();